Refactor routes and add navigation
