+++ web/app/routes/_app.quran.$chapterNumber._index.tsx
@@ -1,4 +1,4 @@
-import type { Route } from '.react-router/types/app/routes/+types/quran.$chapterNumber._index';
+import type { Route } from '.react-router/types/app/routes/+types/_app.quran.$chapterNumber._index';
 import { useSuspenseQuery } from '@tanstack/react-query';
 import { CircleChevronLeft, CircleChevronRight } from 'lucide-react';
 import { Fragment } from 'react';
